refactor(am-tui): split update.rs into focused submodules under update/
Moves all handler logic from the monolithic update.rs and delegation.rs
into focused modules: navigation, selection, text_input, transfer,
profile_actions, trust, and delegation — each under update/.
